+++ docs/changelog.rst
@@ -4,7 +4,8 @@ Changelog
 0.0.2 (Unreleased)
 ------------------
 
--  feat: Log a debug message on :meth:`~pika.clients.Publisher.publish`.
+-  feat: Add :meth:`pika.methods.publish` to publish messages from the context of a consumer callback.
+-  feat: Log a debug message on :meth:`pika.clients.Publisher.publish`.
 
 0.0.1 (2021-11-19)
 ------------------